Move the S60 reporting entity to the virtual platform and pin divisors in the test
It is device-specific cluster setup like the other manufacturer virtual
entities, not tied to the sensor platform. Note in the docstring that it has
to be removed when the quirk declares its own override, since overriding
configs merge with each other, and assert the divisors the raw values assume.
