+++ tests/test_sensor.py
@@ -2193,17 +2193,23 @@ async def test_em_poller_runs_independently_of_entity_enabled_state(
 
 
 @pytest.mark.parametrize(
-    ("device_file", "expected_voltage_change", "expected_current_change"),
+    (
+        "device_file",
+        "divisors",
+        "expected_voltage_change",
+        "expected_current_change",
+    ),
     [
-        # SONOFF S60 reports with a divisor of 100: 1 V / 0.05 A
-        ("sonoff-s60zbtpg-0x00001002.json", 100, 5),
+        # SONOFF S60 reports with divisors of 100: raw 100 / 5 are 1 V / 0.05 A
+        ("sonoff-s60zbtpg-0x00001002.json", (100, 100), 100, 5),
         # Other devices keep the default raw reportable change
-        ("shelly-1pm.json", 1, 1),
+        ("shelly-1pm.json", (100, 1000), 1, 1),
     ],
 )
 async def test_sonoff_s60_electrical_measurement_reporting(
     zha_gateway: Gateway,
     device_file: str,
+    divisors: tuple[int, int],
     expected_voltage_change: int,
     expected_current_change: int,
 ) -> None:
@@ -2213,6 +2219,12 @@ async def test_sonoff_s60_electrical_measurement_reporting(
     )
     cluster = zigpy_device.endpoints[1].electrical_measurement
     await join_zigpy_device(zha_gateway, zigpy_device)
+
+    # The raw reportable changes only mean 1 V / 0.05 A for these divisors
+    assert (
+        cluster.get(EMAttrs.ac_voltage_divisor.id),
+        cluster.get(EMAttrs.ac_current_divisor.id),
+    ) == divisors
 
     assert len(cluster.configure_reporting_multiple.mock_calls) == 1
     configured = cluster.configure_reporting_multiple.call_args.args[0]
